@@ -115,7 +115,8 @@
         "symfony/expression-language": "^5.4 || ^6.0",
         "symfony/filesystem": "^5.4 || ^6.0",
         "symfony/finder": "^5.4 || ^6.0",
-        "symfony/framework-bundle": "^5.4 || ^6.0",
+        "symfony/framework-bundle": "^6.1.2",
+        "symfony/html-sanitizer": "^6.1",
         "symfony/http-client": "^5.4 || ^6.0",
         "symfony/http-client-contracts": "^2.4 || ^3.1",
         "symfony/http-foundation": "^5.4 || ^6.0",
@@ -123,7 +124,7 @@
         "symfony/intl": "^5.4 || ^6.0",
         "symfony/lock": "^5.4 || ^6.0",
         "symfony/mailer": "^5.4 || ^6.0",
-        "symfony/messenger": "^5.4 || ^6.0",
+        "symfony/messenger": "^6.1",
         "symfony/mime": "^5.4 || ^6.0",
         "symfony/monolog-bridge": "^5.4 || ^6.0",
         "symfony/password-hasher": "^5.4 || ^6.0",
@@ -137,12 +138,12 @@
         "symfony/security-core": "^6.1",
         "symfony/security-csrf": "^6.0",
         "symfony/security-http": "^6.0",
-        "symfony/service-contracts": "^1.1 || ^2.0",
+        "symfony/service-contracts": "^2.2",
         "symfony/string": "^5.4 || ^6.0",
         "symfony/translation": "^5.4 || ^6.0",
         "symfony/translation-contracts": "^2.3",
-        "symfony/twig-bridge": "^5.4 || ^6.0",
-        "symfony/twig-bundle": "^5.4 || ^6.0",
+        "symfony/twig-bridge": "^6.1",
+        "symfony/twig-bundle": "^6.1.1",
         "symfony/uid": "^5.4 || ^6.0",
         "symfony/var-dumper": "^5.4 || ^6.0",
         "symfony/yaml": "^5.4 || ^6.0",
