@@ -10,7 +10,7 @@
         }
 
         // Check if this row has all necessary elements
-        if (8 !== row.querySelectorAll('select, button, a.module_link, img.module_image').length) {
+        if (9 !== row.querySelectorAll('select, button, a.module_link, img.module_image').length) {
             return;
         }
 
@@ -44,15 +44,30 @@
                     case 'copy':
                         bt.addEventListener('click', () => {
                             Backend.getScrollOffset();
+
                             const ntr = tr.cloneNode(true);
                             const selects = tr.querySelectorAll('select');
                             const nselects = ntr.querySelectorAll('select');
+
                             for (let j=0; j<selects.length; j++) {
                                 nselects[j].value = selects[j].value;
                             }
+
+                            ntr.querySelectorAll('[data-original-title]').forEach((el) => {
+                                el.setAttribute('title', el.getAttribute('data-original-title'));
+                                el.removeAttribute('data-original-title');
+                            });
+
+                            initializedRows.set(ntr, true);
                             tr.parentNode.insertBefore(ntr, tr.nextSibling);
+
+                            // Remove the ID of the select before initializing Chosen
+                            const select = ntr.querySelector('select.tl_select')
+                            select.removeAttribute('id');
+
                             ntr.querySelector('.chzn-container').remove();
-                            new Chosen(ntr.querySelector('select.tl_select'));
+                            new Chosen(select);
+
                             addEventsTo(ntr);
                             makeSortable(tbody);
                         });
@@ -61,6 +76,7 @@
                     case 'delete':
                         bt.addEventListener('click', () => {
                             Backend.getScrollOffset();
+
                             if (tbody.children.length > 1) {
                                 tr.remove();
                             } else {
@@ -69,6 +85,7 @@
                                     select.value = select.children[0].value;
                                 });
                             }
+
                             makeSortable(tbody);
                         });
                         break;
@@ -76,7 +93,9 @@
                     case 'enable':
                         bt.addEventListener('click', function() {
                             Backend.getScrollOffset();
+
                             const cbx = bt.previousElementSibling;
+
                             if (cbx.checked) {
                                 cbx.checked = '';
                             } else {
@@ -90,20 +109,24 @@
                             bt.addEventListener('keydown', (event) => {
                                 if (event.code === 'ArrowUp' || event.keyCode === 38) {
                                     event.preventDefault();
+
                                     if (tr.previousElementSibling) {
                                         tr.previousElementSibling.insertAdjacentElement('beforebegin', tr);
                                     } else {
                                         tbody.insertAdjacentElement('beforeend', tr);
                                     }
+
                                     bt.focus();
                                     makeSortable(tbody);
                                 } else if (event.code === 'ArrowDown' || event.keyCode === 40) {
                                     event.preventDefault();
+
                                     if (tr.nextElementSibling) {
                                         tr.nextElementSibling.insertAdjacentElement('afterend', tr);
                                     } else {
                                         tbody.insertAdjacentElement('afterbegin', tr);
                                     }
+
                                     bt.focus();
                                     makeSortable(tbody);
                                 }
@@ -120,17 +143,23 @@
             }
 
             const link = tr.querySelector('a.module_link');
-            const image = tr.querySelector('img.module_image');
+            const images = tr.querySelectorAll('img.module_image');
 
             const updateLink = () => {
                 link.href = link.href.replace(/id=[0-9]+/, 'id=' + select.value);
 
                 if (select.value > 0) {
-                    link.style.display = null;
-                    image.style.display = 'none';
+                    link.classList.remove('hidden');
+
+                    images.forEach((image) => {
+                        image.classList.add('hidden');
+                    });
                 } else {
-                    link.style.display = 'none';
-                    image.style.display = null;
+                    link.classList.add('hidden');
+
+                    images.forEach((image) => {
+                        image.classList.remove('hidden');
+                    });
                 }
             };
 
