@@ -105,7 +105,8 @@ class Configuration implements ConfigurationInterface
                     ->end()
                 ->end()
                 ->scalarNode('console_path')
-                    ->info('The path to the Symfony console.')
+                    ->info('The path to the Symfony console. Defaults to %kernel.project_dir%/bin/console.')
+                    ->cannotBeEmpty()
                     ->defaultValue('%kernel.project_dir%/bin/console')
                 ->end()
                 ->append($this->addMessengerNode())
