@@ -26,7 +26,7 @@ class MessengerCronTest extends TestCase
 {
     public function testIsSkippedIfNotOnCli(): void
     {
-        $cron = new MessengerCron(new Container(), new ProcessUtil(), []);
+        $cron = new MessengerCron(new Container(), new ProcessUtil('bin/console'), []);
 
         $this->expectException(CronExecutionSkippedException::class);
 
@@ -45,6 +45,7 @@ class MessengerCronTest extends TestCase
         $processUtil = $this
             ->getMockBuilder(ProcessUtil::class)
             ->onlyMethods(['createPromise'])
+            ->setConstructorArgs(['bin/console'])
             ->getMock()
         ;
 
